+++ src/Exceptions/OnOfficeException.php
@@ -5,5 +5,31 @@ declare(strict_types=1);
 namespace Katalam\OnOfficeAdapter\Exceptions;
 
 use Exception;
+use Katalam\OnOfficeAdapter\Enums\OnOfficeError;
+use Throwable;
 
-class OnOfficeException extends Exception {}
+class OnOfficeException extends Exception
+{
+    private bool $isResponseError;
+
+    public function __construct(string $message = '', int $code = 0, ?Throwable $previous = null, bool $isResponseError = false)
+    {
+        parent::__construct($message, $code, $previous);
+
+        $this->isResponseError = $isResponseError;
+    }
+
+    public function isResponseError(): bool
+    {
+        return $this->isResponseError;
+    }
+
+    public function getError(): OnOfficeError
+    {
+        if (! $this->isResponseError) {
+            return OnOfficeError::UNKNOWN;
+        }
+
+        return OnOfficeError::tryFrom($this->getCode()) ?? OnOfficeError::UNKNOWN;
+    }
+}