+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 ## main
 - Added error code to on office exception
+- Added a method to get the error from the on office exception
 
 ## v0.9.1
 - Added the api claim back to the request factory